user-service/requirements.txt
Flask==2.3.3
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
flask-jwt-extended==4.5.2
flask-cors==4.0.0
pymysql==1.1.0
python-dotenv==1.0.1
bcrypt==4.0.1
werkzeug==2.3.7
gunicorn
cryptography
flask-graphql
graphene
graphene-sqlalchemy
